+++ frontend/src/lib/components/LemonLabel/LemonLabel.tsx
@@ -2,15 +2,16 @@ import './LemonLabel.scss'
 import React from 'react'
 import { Tooltip } from '../Tooltip'
 import { IconInfo } from '../icons'
+import clsx from 'clsx'
 
 export interface LemonLabelProps
-    extends Pick<React.LabelHTMLAttributes<HTMLLabelElement>, 'htmlFor' | 'form' | 'children'> {
+    extends Pick<React.LabelHTMLAttributes<HTMLLabelElement>, 'htmlFor' | 'form' | 'children' | 'className'> {
     info?: React.ReactNode
 }
 
-export function LemonLabel({ children, info, ...props }: LemonLabelProps): JSX.Element {
+export function LemonLabel({ children, info, className, ...props }: LemonLabelProps): JSX.Element {
     return (
-        <label className={'LemonLabel'} {...props}>
+        <label className={clsx('LemonLabel', className)} {...props}>
             {children}
 
             {info ? (
