+++ frontend/src/lib/components/Subscriptions/views/EditSubscription.tsx
@@ -4,7 +4,7 @@ import { LemonButton } from 'lib/components/LemonButton'
 import { VerticalForm } from 'lib/forms/VerticalForm'
 import { membersLogic } from 'scenes/organization/Settings/membersLogic'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
-import { Field } from 'lib/forms/Field'
+import { Field } from 'lib/forms/FieldV2'
 import { dayjs } from 'lib/dayjs'
 import { LemonSelect, LemonSelectOptions, LemonSelectProps } from 'lib/components/LemonSelect'
 import { subscriptionLogic } from '../subscriptionLogic'
@@ -32,6 +32,8 @@ import { usersLemonSelectOptions } from 'lib/components/UserSelectItem'
 import { integrationsLogic } from 'scenes/project/Settings/integrationsLogic'
 import { urls } from 'scenes/urls'
 import { Skeleton } from 'antd'
+import { Form } from 'kea-forms'
+import { LemonLabel } from 'lib/components/LemonLabel/LemonLabel'
 
 interface EditSubscriptionProps extends SubscriptionBaseProps {
     id: number | 'new'
@@ -99,98 +101,102 @@ export function EditSubscription({
 
     return (
         <>
-            <VerticalForm logic={subscriptionLogic} props={logicProps} formKey="subscription" enableFormOnSubmit>
-                <header className="flex items-center border-b pb-2">
-                    <LemonButton status="stealth" onClick={onCancel} size="small">
-                        <IconChevronLeft fontSize={'1rem'} />
-                        Back
-                    </LemonButton>
-                    <LemonDivider vertical />
+            <header className="flex items-center border-b pb-2">
+                <LemonButton status="stealth" onClick={onCancel} size="small">
+                    <IconChevronLeft fontSize={'1rem'} />
+                    Back
+                </LemonButton>
+                <LemonDivider vertical />
 
-                    <h4 className="mt-2">{id === 'new' ? 'New' : 'Edit '} Subscription</h4>
-                </header>
+                <h4 className="mt-2">{id === 'new' ? 'New' : 'Edit '} Subscription</h4>
+            </header>
 
-                {!subscription ? (
-                    subscriptionLoading ? (
-                        <>
-                            <Skeleton />
-                            <Skeleton />
-                            <Skeleton />
-                        </>
-                    ) : (
-                        <section className="p-4 text-center">
-                            <h2>Not found</h2>
-                            <p>This subscription could not be found. It may have been deleted.</p>
-                        </section>
-                    )
-                ) : (
-                    <section>
-                        {subscription?.created_by ? (
-                            <UserActivityIndicator
-                                at={subscription.created_at}
-                                by={subscription.created_by}
-                                prefix={'Created'}
-                                className={'mb-4'}
-                            />
-                        ) : null}
-
-                        {siteUrlMisconfigured && (
-                            <AlertMessage type="warning">
-                                <>
-                                    Your <code>SITE_URL</code> environment variable seems misconfigured. Your{' '}
-                                    <code>SITE_URL</code> is set to{' '}
-                                    <b>
-                                        <code>{preflight?.site_url}</code>
-                                    </b>{' '}
-                                    but you're currently browsing this page from{' '}
-                                    <b>
-                                        <code>{window.location.origin}</code>
-                                    </b>
-                                    . <br />
-                                    If this value is not configured correctly PostHog may be unable to correctly send
-                                    Subscriptions.{' '}
-                                    <a
-                                        target="_blank"
-                                        rel="noopener"
-                                        href="https://posthog.com/docs/configuring-posthog/environment-variables?utm_medium=in-product&utm_campaign=subcriptions-system-status-site-url-misconfig"
-                                    >
-                                        Learn more <IconOpenInNew />
-                                    </a>
-                                </>
-                            </AlertMessage>
-                        )}
-
-                        <Field name={'title'} label={'Name'}>
-                            <LemonInput placeholder="e.g. Weekly team report" />
-                        </Field>
-
-                        <Field name={'target_type'} label={'Destination'}>
-                            <LemonSelect options={targetTypeOptions} {...commonSelectProps} />
-                        </Field>
-
-                        {subscription.target_type === 'email' ? (
+            <Form logic={subscriptionLogic} props={logicProps} formKey="subscription" enableFormOnSubmit>
+                <div className="space-y-4 my-4">
+                    {!subscription ? (
+                        subscriptionLoading ? (
                             <>
-                                {emailDisabled && (
-                                    <AlertMessage type="error">
-                                        <>
-                                            Email subscriptions are not currently possible as this PostHog instance
-                                            isn't{' '}
-                                            <a
-                                                href="https://posthog.com/docs/self-host/configure/email"
-                                                target="_blank"
-                                                rel="noopener"
-                                            >
-                                                configured&nbsp;to&nbsp;send&nbsp;emails&nbsp;
-                                                <IconOpenInNew />
-                                            </a>
-                                            .
-                                        </>
-                                    </AlertMessage>
-                                )}
+                                <Skeleton />
+                                <Skeleton />
+                                <Skeleton />
+                            </>
+                        ) : (
+                            <section className="p-4 text-center">
+                                <h2>Not found</h2>
+                                <p>This subscription could not be found. It may have been deleted.</p>
+                            </section>
+                        )
+                    ) : (
+                        <section>
+                            {subscription?.created_by ? (
+                                <UserActivityIndicator
+                                    at={subscription.created_at}
+                                    by={subscription.created_by}
+                                    prefix={'Created'}
+                                    className={'mb-4'}
+                                />
+                            ) : null}
 
-                                <Field name={'target_value'} label={'Who do you want to subscribe'}>
-                                    {({ value, onChange }) => (
-                                        <>
+                            {siteUrlMisconfigured && (
+                                <AlertMessage type="warning">
+                                    <>
+                                        Your <code>SITE_URL</code> environment variable seems misconfigured. Your{' '}
+                                        <code>SITE_URL</code> is set to{' '}
+                                        <b>
+                                            <code>{preflight?.site_url}</code>
+                                        </b>{' '}
+                                        but you're currently browsing this page from{' '}
+                                        <b>
+                                            <code>{window.location.origin}</code>
+                                        </b>
+                                        . <br />
+                                        If this value is not configured correctly PostHog may be unable to correctly
+                                        send Subscriptions.{' '}
+                                        <a
+                                            target="_blank"
+                                            rel="noopener"
+                                            href="https://posthog.com/docs/configuring-posthog/environment-variables?utm_medium=in-product&utm_campaign=subcriptions-system-status-site-url-misconfig"
+                                        >
+                                            Learn more <IconOpenInNew />
+                                        </a>
+                                    </>
+                                </AlertMessage>
+                            )}
+
+                            <Field name={'title'} label={'Name'}>
+                                <LemonInput placeholder="e.g. Weekly team report" />
+                            </Field>
+
+                            <Field name={'target_type'} label={'Destination'}>
+                                <LemonSelect options={targetTypeOptions} {...commonSelectProps} />
+                            </Field>
+
+                            {subscription.target_type === 'email' ? (
+                                <>
+                                    {emailDisabled && (
+                                        <AlertMessage type="error">
+                                            <>
+                                                Email subscriptions are not currently possible as this PostHog instance
+                                                isn't{' '}
+                                                <a
+                                                    href="https://posthog.com/docs/self-host/configure/email"
+                                                    target="_blank"
+                                                    rel="noopener"
+                                                >
+                                                    configured&nbsp;to&nbsp;send&nbsp;emails&nbsp;
+                                                    <IconOpenInNew />
+                                                </a>
+                                                .
+                                            </>
+                                        </AlertMessage>
+                                    )}
+
+                                    <Field
+                                        name={'target_value'}
+                                        label={'Who do you want to subscribe'}
+                                        help={'Enter the email addresses of the users you want to share with'}
+                                    >
+                                        {({ value, onChange }) => (
                                             <LemonSelectMultiple
                                                 onChange={(val) => onChange(val.join(','))}
                                                 value={value?.split(',').filter(Boolean)}
@@ -201,76 +207,64 @@ export function EditSubscription({
                                                 loading={membersLoading}
                                                 placeholder="Enter an email address"
                                             />
-                                            <div className="text-xs text-muted mt-2">
-                                                Enter the email addresses of the users you want to share with
-                                            </div>
-                                        </>
-                                    )}
-                                </Field>
-
-                                <Field name={'invite_message'} label={'Message (optional)'}>
-                                    <LemonTextArea placeholder="Your message to new subscribers (optional)" />
-                                </Field>
-                            </>
-                        ) : null}
-
-                        {subscription.target_type === 'slack' ? (
-                            <>
-                                {slackDisabled ? (
-                                    <>
-                                        {addToSlackButtonUrl() ? (
-                                            <AlertMessage type="info">
-                                                <div className="flex justify-between gap-2">
-                                                    <span>
-                                                        Slack is not yet configured for this project. Add PostHog to
-                                                        your Slack workspace to continue.
-                                                    </span>
-                                                    <a
-                                                        href={
-                                                            addToSlackButtonUrl(
-                                                                window.location.pathname + '?target_type=slack'
-                                                            ) || ''
-                                                        }
-                                                    >
-                                                        <img
-                                                            alt="Add to Slack"
-                                                            height="40"
-                                                            width="139"
-                                                            src="https://platform.slack-edge.com/img/add_to_slack.png"
-                                                            srcSet="https://platform.slack-edge.com/img/add_to_slack.png 1x, https://platform.slack-edge.com/img/add_to_slack@2x.png 2x"
-                                                        />
-                                                    </a>
-                                                </div>
-                                            </AlertMessage>
-                                        ) : (
-                                            <AlertMessage type="error">
-                                                <>
-                                                    Slack is not yet configured for this project. You can configure it
-                                                    at{' '}
-                                                    <Link to={`${urls.projectSettings()}#slack`}>
-                                                        {' '}
-                                                        Slack Integration settings
-                                                    </Link>
-                                                    .
-                                                </>
-                                            </AlertMessage>
                                         )}
-                                    </>
-                                ) : (
-                                    <>
-                                        <Field name={'target_value'} label={'Which Slack channel to send reports to'}>
-                                            {({ value, onChange }) => (
-                                                <>
-                                                    <LemonSelectMultiple
-                                                        onChange={(val) => onChange(val)}
-                                                        value={value}
-                                                        disabled={slackDisabled}
-                                                        mode="single"
-                                                        data-attr="select-slack-channel"
-                                                        options={slackChannelOptions}
-                                                        loading={slackChannelsLoading}
-                                                    />
-                                                    <div className="text-xs text-muted mt-2">
+                                    </Field>
+
+                                    <Field name={'invite_message'} label={'Message (optional)'}>
+                                        <LemonTextArea placeholder="Your message to new subscribers (optional)" />
+                                    </Field>
+                                </>
+                            ) : null}
+
+                            {subscription.target_type === 'slack' ? (
+                                <>
+                                    {slackDisabled ? (
+                                        <>
+                                            {addToSlackButtonUrl() ? (
+                                                <AlertMessage type="info">
+                                                    <div className="flex justify-between gap-2">
+                                                        <span>
+                                                            Slack is not yet configured for this project. Add PostHog to
+                                                            your Slack workspace to continue.
+                                                        </span>
+                                                        <a
+                                                            href={
+                                                                addToSlackButtonUrl(
+                                                                    window.location.pathname + '?target_type=slack'
+                                                                ) || ''
+                                                            }
+                                                        >
+                                                            <img
+                                                                alt="Add to Slack"
+                                                                height="40"
+                                                                width="139"
+                                                                src="https://platform.slack-edge.com/img/add_to_slack.png"
+                                                                srcSet="https://platform.slack-edge.com/img/add_to_slack.png 1x, https://platform.slack-edge.com/img/add_to_slack@2x.png 2x"
+                                                            />
+                                                        </a>
+                                                    </div>
+                                                </AlertMessage>
+                                            ) : (
+                                                <AlertMessage type="error">
+                                                    <>
+                                                        Slack is not yet configured for this project. You can configure
+                                                        it at{' '}
+                                                        <Link to={`${urls.projectSettings()}#slack`}>
+                                                            {' '}
+                                                            Slack Integration settings
+                                                        </Link>
+                                                        .
+                                                    </>
+                                                </AlertMessage>
+                                            )}
+                                        </>
+                                    ) : (
+                                        <>
+                                            <Field
+                                                name={'target_value'}
+                                                label={'Which Slack channel to send reports to'}
+                                                help={
+                                                    <>
                                                         Private channels are only shown if you have{' '}
                                                         <a
                                                             href="https://posthog.com/docs/integrate/third-party/slack"
@@ -280,140 +274,152 @@ export function EditSubscription({
                                                             added the PostHog Slack App
                                                         </a>{' '}
                                                         to them
-                                                    </div>
-                                                </>
-                                            )}
-                                        </Field>
-
-                                        {showSlackMembershipWarning ? (
-                                            <Field name={'memberOfSlackChannel'}>
-                                                <AlertMessage type="info">
-                                                    <div className="flex gap-2 items-center">
-                                                        <span>
-                                                            The PostHog Slack App is not in this channel. Please add it
-                                                            to the channel otherwise Subscriptions will fail to be
-                                                            delivered.{' '}
-                                                            <a
-                                                                href="https://posthog.com/docs/integrate/third-party/slack"
-                                                                target="_blank"
-                                                                rel="noopener"
-                                                            >
-                                                                See the Docs for more information
-                                                            </a>
-                                                        </span>
-                                                        <LemonButton
-                                                            type="secondary"
-                                                            onClick={() => loadSlackChannels()}
-                                                            loading={slackChannelsLoading}
-                                                        >
-                                                            Check again
-                                                        </LemonButton>
-                                                    </div>
-                                                </AlertMessage>
+                                                    </>
+                                                }
+                                            >
+                                                {({ value, onChange }) => (
+                                                    <LemonSelectMultiple
+                                                        onChange={(val) => onChange(val)}
+                                                        value={value}
+                                                        disabled={slackDisabled}
+                                                        mode="single"
+                                                        data-attr="select-slack-channel"
+                                                        options={slackChannelOptions}
+                                                        loading={slackChannelsLoading}
+                                                    />
+                                                )}
                                             </Field>
-                                        ) : null}
-                                    </>
-                                )}
-                            </>
-                        ) : null}
 
-                        {subscription.target_type === 'webhook' ? (
-                            <>
-                                <Field name={'target_value'} label={'Webhook URL'}>
-                                    <LemonInput placeholder="https://example.com/webhooks/1234" />
-                                </Field>
-                                <div className="text-xs text-muted mt-2">
-                                    Webhooks will be called with a HTTP POST request. The webhook endpoint should
-                                    respond with a healthy HTTP code (2xx).
-                                </div>
-                            </>
-                        ) : null}
-
-                        <div>
-                            <div className="ant-form-item-label">
-                                <label title="Recurrence">Recurrence</label>
-                            </div>
-                            <div className="flex gap-2 items-center rounded border p-2 flex-wrap">
-                                <span>Send every</span>
-                                <Field name={'interval'} style={{ marginBottom: 0 }}>
-                                    <LemonSelect {...commonSelectProps} options={intervalOptions} />
-                                </Field>
-                                <Field name={'frequency'} style={{ marginBottom: 0 }}>
-                                    <LemonSelect {...commonSelectProps} options={frequencyOptions} />
-                                </Field>
-
-                                {subscription.frequency === 'weekly' && (
-                                    <>
-                                        <span>on</span>
-                                        <Field name={'byweekday'} style={{ marginBottom: 0 }}>
-                                            {({ value, onChange }) => (
-                                                <LemonSelect
-                                                    {...commonSelectProps}
-                                                    options={weekdayOptions}
-                                                    value={value ? value[0] : null}
-                                                    onChange={(val) => onChange([val])}
-                                                />
-                                            )}
-                                        </Field>
-                                    </>
-                                )}
-
-                                {subscription.frequency === 'monthly' && (
-                                    <>
-                                        <span>on the</span>
-                                        <Field name={'bysetpos'} style={{ marginBottom: 0 }}>
-                                            {({ value, onChange }) => (
-                                                <LemonSelect
-                                                    {...commonSelectProps}
-                                                    options={bysetposOptions}
-                                                    value={value ? String(value) : null}
-                                                    onChange={(val) => {
-                                                        onChange(typeof val === 'string' ? parseInt(val, 10) : null)
-                                                    }}
-                                                />
-                                            )}
-                                        </Field>
-                                        <Field name={'byweekday'} style={{ marginBottom: 0 }}>
-                                            {({ value, onChange }) => (
-                                                <LemonSelect
-                                                    {...commonSelectProps}
-                                                    dropdownMatchSelectWidth={false}
-                                                    options={monthlyWeekdayOptions}
-                                                    // "day" is a special case where it is a list of all available days
-                                                    value={value ? (value.length === 1 ? value[0] : 'day') : null}
-                                                    onChange={(val) =>
-                                                        onChange(val === 'day' ? Object.keys(weekdayOptions) : [val])
-                                                    }
-                                                />
-                                            )}
-                                        </Field>
-                                    </>
-                                )}
-                                <span>by</span>
-                                <Field name={'start_date'}>
-                                    {({ value, onChange }) => (
-                                        <LemonSelect
-                                            {...commonSelectProps}
-                                            options={timeOptions}
-                                            value={dayjs(value).hour().toString()}
-                                            onChange={(val) => {
-                                                onChange(
-                                                    dayjs()
-                                                        .hour(typeof val === 'string' ? parseInt(val, 10) : 0)
-                                                        .minute(0)
-                                                        .second(0)
-                                                        .toISOString()
-                                                )
-                                            }}
-                                        />
+                                            {showSlackMembershipWarning ? (
+                                                <Field name={'memberOfSlackChannel'}>
+                                                    <AlertMessage type="info">
+                                                        <div className="flex gap-2 items-center">
+                                                            <span>
+                                                                The PostHog Slack App is not in this channel. Please add
+                                                                it to the channel otherwise Subscriptions will fail to
+                                                                be delivered.{' '}
+                                                                <a
+                                                                    href="https://posthog.com/docs/integrate/third-party/slack"
+                                                                    target="_blank"
+                                                                    rel="noopener"
+                                                                >
+                                                                    See the Docs for more information
+                                                                </a>
+                                                            </span>
+                                                            <LemonButton
+                                                                type="secondary"
+                                                                onClick={() => loadSlackChannels()}
+                                                                loading={slackChannelsLoading}
+                                                            >
+                                                                Check again
+                                                            </LemonButton>
+                                                        </div>
+                                                    </AlertMessage>
+                                                </Field>
+                                            ) : null}
+                                        </>
                                     )}
-                                </Field>
-                            </div>
-                        </div>
-                    </section>
-                )}
+                                </>
+                            ) : null}
 
-                <footer className="flex justify-between pt-4">
+                            {subscription.target_type === 'webhook' ? (
+                                <>
+                                    <Field name={'target_value'} label={'Webhook URL'}>
+                                        <LemonInput placeholder="https://example.com/webhooks/1234" />
+                                    </Field>
+                                    <div className="text-xs text-muted mt-2">
+                                        Webhooks will be called with a HTTP POST request. The webhook endpoint should
+                                        respond with a healthy HTTP code (2xx).
+                                    </div>
+                                </>
+                            ) : null}
+
+                            <div>
+                                <LemonLabel className="mb-2">Recurrence</LemonLabel>
+                                <div className="flex gap-2 items-center rounded border p-2 flex-wrap">
+                                    <span>Send every</span>
+                                    <Field name={'interval'}>
+                                        <LemonSelect {...commonSelectProps} options={intervalOptions} />
+                                    </Field>
+                                    <Field name={'frequency'}>
+                                        <LemonSelect {...commonSelectProps} options={frequencyOptions} />
+                                    </Field>
+
+                                    {subscription.frequency === 'weekly' && (
+                                        <>
+                                            <span>on</span>
+                                            <Field name={'byweekday'}>
+                                                {({ value, onChange }) => (
+                                                    <LemonSelect
+                                                        {...commonSelectProps}
+                                                        options={weekdayOptions}
+                                                        value={value ? value[0] : null}
+                                                        onChange={(val) => onChange([val])}
+                                                    />
+                                                )}
+                                            </Field>
+                                        </>
+                                    )}
+
+                                    {subscription.frequency === 'monthly' && (
+                                        <>
+                                            <span>on the</span>
+                                            <Field name={'bysetpos'}>
+                                                {({ value, onChange }) => (
+                                                    <LemonSelect
+                                                        {...commonSelectProps}
+                                                        options={bysetposOptions}
+                                                        value={value ? String(value) : null}
+                                                        onChange={(val) => {
+                                                            onChange(typeof val === 'string' ? parseInt(val, 10) : null)
+                                                        }}
+                                                    />
+                                                )}
+                                            </Field>
+                                            <Field name={'byweekday'}>
+                                                {({ value, onChange }) => (
+                                                    <LemonSelect
+                                                        {...commonSelectProps}
+                                                        dropdownMatchSelectWidth={false}
+                                                        options={monthlyWeekdayOptions}
+                                                        // "day" is a special case where it is a list of all available days
+                                                        value={value ? (value.length === 1 ? value[0] : 'day') : null}
+                                                        onChange={(val) =>
+                                                            onChange(
+                                                                val === 'day' ? Object.keys(weekdayOptions) : [val]
+                                                            )
+                                                        }
+                                                    />
+                                                )}
+                                            </Field>
+                                        </>
+                                    )}
+                                    <span>by</span>
+                                    <Field name={'start_date'}>
+                                        {({ value, onChange }) => (
+                                            <LemonSelect
+                                                {...commonSelectProps}
+                                                options={timeOptions}
+                                                value={dayjs(value).hour().toString()}
+                                                onChange={(val) => {
+                                                    onChange(
+                                                        dayjs()
+                                                            .hour(typeof val === 'string' ? parseInt(val, 10) : 0)
+                                                            .minute(0)
+                                                            .second(0)
+                                                            .toISOString()
+                                                    )
+                                                }}
+                                            />
+                                        )}
+                                    </Field>
+                                </div>
+                            </div>
+                        </section>
+                    )}
+                </div>
+
+                <footer className="flex justify-between pt-4 border-t">
                     <div>
                         {subscription && id !== 'new' && (
                             <LemonButton
@@ -440,7 +446,7 @@ export function EditSubscription({
                         </LemonButton>
                     </div>
                 </footer>
-            </VerticalForm>
+            </Form>
         </>
     )
 }
