+++ frontend/src/lib/forms/Field.stories.tsx
@@ -1,44 +1,70 @@
 import React from 'react'
-import { kea, path, useAllValues } from 'kea'
 import { ComponentMeta } from '@storybook/react'
-import { VerticalForm } from 'lib/forms/VerticalForm'
-import { Field, FieldProps } from './Field'
+import { Field, PureField } from './FieldV2'
+import { LemonButton, LemonCheckbox, LemonInput, LemonSelect, LemonTextArea } from '@posthog/lemon-ui'
+import { kea, path, useAllValues } from 'kea'
+import { Form, forms } from 'kea-forms'
+
 import type { formLogicType } from './Field.storiesType'
-import { capitalizeFirstLetter } from 'lib/utils'
-import { forms } from 'kea-forms'
-import { LemonInput } from 'lib/components/LemonInput/LemonInput'
 
 export default {
     title: 'Forms/Field',
-    component: Field,
-    argTypes: {
-        name: {
-            control: { type: 'text' },
-            description: 'Key of field in the form',
-            type: 'string',
-        },
-        label: {
-            control: { type: 'text' },
-            description: 'Text that appears above the field',
-            type: 'string',
-        },
-        hint: {
-            control: { type: 'text' },
-            description: 'Text that appears below the field',
-            type: 'string',
-        },
-        noStyle: {
-            control: { type: 'boolean' },
-            description: 'Directly return `children` without labels, errors, etc.',
-        },
-        children: {
-            description: 'The component to render',
-        },
-        template: {
-            description: 'Form template to use (WIP)',
+    component: PureField,
+    parameters: {
+        docs: {
+            description: {
+                component: `
+
+[Related Figma area](https://www.figma.com/file/Y9G24U4r04nEjIDGIEGuKI/PostHog-Design-System-One?node-id=3139%3A1388)
+
+Fields are a wrapping component that take care of rendering a label, input and error messages in a standard format.
+
+They can be used in a kea-forms controlled way via \`Field\` or a pure way via \`PureField\`.
+`,
+            },
         },
     },
-} as ComponentMeta<typeof Field>
+} as ComponentMeta<typeof PureField>
+
+export const _PureFields = (): JSX.Element => {
+    return (
+        <div className="space-y-4">
+            <PureField
+                label={
+                    <>
+                        Text input label <span>(Optional)</span>
+                    </>
+                }
+                hint={
+                    <>
+                        Optional descriptive or supportive text for the preceeding form element. This content can wrap
+                        to multiple lines.
+                    </>
+                }
+            >
+                <LemonInput placeholder="Optional descriptive placeholder text" />
+            </PureField>
+
+            <PureField label={'Select label'} info={<>With info!</>}>
+                <LemonSelect options={{ foo: { label: 'bar' } }} fullWidth />
+            </PureField>
+
+            <PureField label="Textarea label" error="This field has an error">
+                <LemonTextArea />
+            </PureField>
+            <PureField>
+                <LemonCheckbox bordered label="Checkbox labels are set differently" fullWidth />
+            </PureField>
+
+            <div className="flex justify-end gap-2 border-t mt-4 pt-4">
+                <LemonButton type="secondary">Cancel</LemonButton>
+                <LemonButton htmlType="submit" type="primary">
+                    Submit
+                </LemonButton>
+            </div>
+        </div>
+    )
+}
 
 const formLogic = kea<formLogicType>([
     path(['lib', 'forms', 'Field', 'stories']),
@@ -49,9 +75,10 @@ const formLogic = kea<formLogicType>([
                 email: '',
                 pineappleOnPizza: false,
             },
-            errors: ({ name, email }) => ({
+            errors: ({ name, email, pineappleOnPizza }) => ({
                 name: !name ? 'Please enter your name' : null,
                 email: !email ? 'Please enter your email' : !email.includes('@') ? 'not a valid email' : null,
+                pineappleOnPizza: pineappleOnPizza ? 'I think you meant to leave this unchecked...' : null,
             }),
             submit: async (_, breakpoint) => {
                 await breakpoint(3000)
@@ -82,57 +109,52 @@ function useSpecificFormValues(formKey: string): Record<string, any> {
     )
 }
 
-export function Field_(props: FieldProps): JSX.Element {
+export const _FieldsWithKeaForm = (): JSX.Element => {
     const formKey = 'myForm'
     const formValues = useSpecificFormValues(formKey)
+
     return (
-        <VerticalForm logic={formLogic} formKey={formKey} enableFormOnSubmit>
-            <Field {...props} name={props.name || 'name'} label={props.label || 'Name'}>
-                <LemonInput />
-            </Field>
+        <Form logic={formLogic} formKey={formKey} enableFormOnSubmit>
+            <div className="space-y-4">
+                <Field
+                    name="name"
+                    label={
+                        <>
+                            Name <span>(What should we call you?)</span>
+                        </>
+                    }
+                    hint={
+                        <>
+                            Optional descriptive or supportive text for the preceeding form element. This content can
+                            wrap to multiple lines.
+                        </>
+                    }
+                >
+                    <LemonInput placeholder="Jon Snow" />
+                </Field>
 
-            <Field {...props} name={props.name || 'email'} label={props.label || 'Email'}>
-                <LemonInput />
-            </Field>
+                <Field name="select" label={'Select label'} info={<>With info!</>}>
+                    <LemonSelect options={{ foo: { label: 'bar' } }} fullWidth />
+                </Field>
 
-            <Field
-                {...props}
-                showOptional={props.showOptional ?? true}
-                name={props.name || 'pineappleOnPizza'}
-                label={props.label || 'Pineapple on pizza preference'}
-            >
-                {({ value, onChange }) => (
-                    <input type="checkbox" checked={value} onChange={(e) => onChange(e.target.checked)} />
-                )}
-            </Field>
-            <input
-                type="submit"
-                value={formValues[`is${capitalizeFirstLetter(formKey)}Submitting`] ? 'Submitting...' : 'Submit'}
-                disabled={formValues[`is${capitalizeFirstLetter(formKey)}Submitting`]}
-            />
-            <pre>
-                {'\n'}formLogic.values = {JSON.stringify(formValues, null, 2)}
-            </pre>
-        </VerticalForm>
-    )
-}
+                <Field name="email" label="Email address">
+                    <LemonInput type="email" />
+                </Field>
+                <Field name="pineappleOnPizza">
+                    <LemonCheckbox bordered label="Checkbox labels are set differently" fullWidth />
+                </Field>
 
-export function TextField(props: FieldProps): JSX.Element {
-    const formKey = 'simpleForm'
-    const formValues = useSpecificFormValues(formKey)
-    return (
-        <VerticalForm logic={formLogic} formKey={formKey} enableFormOnSubmit>
-            <Field {...props} name={props.name || 'name'} label={props.label || 'Name'}>
-                <LemonInput />
-            </Field>
-            <input
-                type="submit"
-                value={formValues[`is${capitalizeFirstLetter(formKey)}Submitting`] ? 'Submitting...' : 'Submit'}
-                disabled={formValues[`is${capitalizeFirstLetter(formKey)}Submitting`]}
-            />
-            <pre>
-                {'\n'}formLogic.values = {JSON.stringify(formValues, null, 2)}
-            </pre>
-        </VerticalForm>
+                <div className="flex justify-end gap-2 border-t mt-4 pt-4">
+                    <LemonButton type="secondary">Cancel</LemonButton>
+                    <LemonButton htmlType="submit" type="primary">
+                        Submit
+                    </LemonButton>
+                </div>
+
+                <pre className="rounded-lg text-white bg-default p-2 m-2">
+                    formLogic.values = {JSON.stringify(formValues, null, 2)}
+                </pre>
+            </div>
+        </Form>
     )
 }
