+++ frontend/src/lib/forms/FieldV2.tsx
@@ -15,12 +15,13 @@ export type PureFieldProps = {
     children?: React.ReactNode
 }
 
+/** A "Pure" field - used when you want the Field styles without the Kea form functionality */
 export const PureField = ({ label, info, error, help, children }: PureFieldProps): JSX.Element => {
     return (
         <div className="flex flex-col gap-2">
             {label ? <LemonLabel info={info}>{label}</LemonLabel> : null}
             {children}
-            {help ? <div className="text-muted">{help}</div> : null}
+            {help ? <div className="text-muted text-xs">{help}</div> : null}
             {error ? (
                 <div className="text-danger flex items-center gap-1">
                     <IconErrorOutline className="text-xl" /> {error}
@@ -30,17 +31,16 @@ export const PureField = ({ label, info, error, help, children }: PureFieldProps
     )
 }
 
-export type FieldProps = Omit<PureFieldProps, 'children' | 'error'> & KeaFieldProps
+export type FieldProps = Omit<PureFieldProps, 'children' | 'error'> & Pick<KeaFieldProps, 'children' | 'name'>
 
-export const Field = ({ name, ...keaFieldProps }: FieldProps): JSX.Element => {
+export const Field = ({ name, help, ...keaFieldProps }: FieldProps): JSX.Element => {
     /** Drop-in replacement antd template for kea forms */
-    const template: FieldProps['template'] = ({ label, kids, error }) => {
-        console.log({ label, kids, error })
+    const template: KeaFieldProps['template'] = ({ label, kids, error }) => {
         return (
-            <PureField label={label} error={error}>
+            <PureField label={label} error={error} help={help}>
                 {kids}
             </PureField>
         )
     }
-    return <KeaField {...keaFieldProps} name={name} template={template} />
+    return <KeaField {...keaFieldProps} name={name} template={template} noStyle />
 }
